Refactoring: move function which is for a specific test case
